Move BicUpdates/PreCAWG/DAWG-PAC sections to external components; add tabs to organize PreCAWG/DAWG-PAC content
