+++ src/MultiOmicsWorkingGroups/bicUpdates.jsx
@@ -1,0 +1,39 @@
+import React from 'react';
+
+function BicUpdates() {
+  return (
+    <div className="multi-omics-working-groups-content-container bic-updates mt-5">
+      <h3 className="border-bottom mb-3 pb-3">BIC Updates</h3>
+      <p>
+        The Bioinformatics Center is committed to keeping our community informed
+        and up-to-date with the latest developments in bioinformatics resources,
+        tools, and data. To this end, we provide{' '}
+        <a
+          href="https://docs.google.com/document/d/11RneqlJsUPcj3XxKuoUhv09z-briz5WytL8gJvzgOPY/edit?usp=sharing"
+          target="_blank"
+          rel="noopener noreferrer"
+        >
+          monthly BIC Updates
+        </a>{' '}
+        that encapsulate all recent advancements, new resources, and crucial
+        data updates facilitated by our center.{' '}
+        <a
+          href="https://drive.google.com/drive/folders/1a5lMTVRHhsWGC7NKKBWGIvByZ8zAiAj5?usp=drive_link"
+          target="_blank"
+          rel="noopener noreferrer"
+        >
+          The archive of all the monthly BIC Updates is available here
+        </a>
+        .
+      </p>
+      <p>
+        Note: BIC Updates are disseminated monthly through email by Jane Lu. To
+        be added to the distribution list, please contact Jane at{' '}
+        <a href="mailto:motrpac-acc@aging.ufl.edu">motrpac-acc@aging.ufl.edu</a>
+        .
+      </p>
+    </div>
+  );
+}
+
+export default BicUpdates;
